debt - move search out of platform
