+++ src/tsconfig.strictNullChecks.json
@@ -338,10 +338,6 @@
 		"./vs/platform/request/electron-main/requestService.ts",
 		"./vs/platform/request/node/request.ts",
 		"./vs/platform/request/node/requestService.ts",
-		"./vs/platform/search/common/replace.ts",
-		"./vs/platform/search/common/search.ts",
-		"./vs/platform/search/test/common/replace.test.ts",
-		"./vs/platform/search/test/common/search.test.ts",
 		"./vs/platform/state/common/state.ts",
 		"./vs/platform/state/node/stateService.ts",
 		"./vs/platform/statusbar/common/statusbar.ts",
@@ -730,6 +726,8 @@
 		"./vs/workbench/services/remote/node/remoteAgentEnvironmentChannel.ts",
 		"./vs/workbench/services/remote/node/remoteAgentService.ts",
 		"./vs/workbench/services/search/common/searchHelpers.ts",
+		"./vs/workbench/services/search/common/replace.ts",
+		"./vs/workbench/services/search/common/search.ts",
 		"./vs/workbench/services/search/node/fileSearch.ts",
 		"./vs/workbench/services/search/node/fileSearchManager.ts",
 		"./vs/workbench/services/search/node/rawSearchService.ts",
@@ -749,6 +747,8 @@
 		"./vs/workbench/services/search/test/node/search.test.ts",
 		"./vs/workbench/services/search/test/node/textSearch.integrationTest.ts",
 		"./vs/workbench/services/search/test/node/textSearchManager.test.ts",
+		"./vs/workbench/services/search/test/common/replace.test.ts",
+		"./vs/workbench/services/search/test/common/search.test.ts",
 		"./vs/workbench/services/textMate/electron-browser/TMGrammars.ts",
 		"./vs/workbench/services/textMate/electron-browser/TMHelper.ts",
 		"./vs/workbench/services/textMate/electron-browser/TMSyntax.ts",
