+++ src/vs/workbench/contrib/search/test/browser/searchActions.test.ts
@@ -14,7 +14,7 @@ import { TestConfigurationService } from 'vs/platform/configuration/test/common/
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { USLayoutResolvedKeybinding } from 'vs/platform/keybinding/common/usLayoutResolvedKeybinding';
-import { IFileMatch } from 'vs/platform/search/common/search';
+import { IFileMatch } from 'vs/workbench/services/search/common/search';
 import { ReplaceAction } from 'vs/workbench/contrib/search/browser/searchActions';
 import { FileMatch, FileMatchOrMatch, Match } from 'vs/workbench/contrib/search/common/searchModel';
 import { MockObjectTree } from 'vs/workbench/contrib/search/test/browser/mockSearchTree';
