+++ src/vs/workbench/services/search/common/search.ts
