+++ src/vs/workbench/services/search/node/ripgrepFileSearch.ts
@@ -11,7 +11,7 @@ import * as objects from 'vs/base/common/objects';
 import * as paths from 'vs/base/common/paths';
 import { isMacintosh as isMac } from 'vs/base/common/platform';
 import * as strings from 'vs/base/common/strings';
-import { IFileQuery, IFolderQuery } from 'vs/platform/search/common/search';
+import { IFileQuery, IFolderQuery } from 'vs/workbench/services/search/common/search';
 import { anchorGlob } from 'vs/workbench/services/search/node/ripgrepSearchUtils';
 import { rgPath } from 'vscode-ripgrep';
