+++ src/vs/workbench/services/search/node/ripgrepSearchUtils.ts
@@ -5,7 +5,7 @@
 
 import { startsWith } from 'vs/base/common/strings';
 import { ILogService } from 'vs/platform/log/common/log';
-import { SearchRange, TextSearchMatch } from 'vs/platform/search/common/search';
+import { SearchRange, TextSearchMatch } from 'vs/workbench/services/search/common/search';
 import * as vscode from 'vscode';
 import { mapArrayOrNot } from 'vs/base/common/arrays';
