+++ src/vs/workbench/services/search/node/searchIpc.ts
@@ -5,7 +5,7 @@
 
 import { Event } from 'vs/base/common/event';
 import { IChannel, IServerChannel } from 'vs/base/parts/ipc/node/ipc';
-import { IRawFileQuery, IRawTextQuery } from 'vs/platform/search/common/search';
+import { IRawFileQuery, IRawTextQuery } from 'vs/workbench/services/search/common/search';
 import { IRawSearchService, ISerializedSearchComplete, ISerializedSearchProgressItem } from './search';
 
 export class SearchChannel implements IServerChannel {
