+++ src/vs/workbench/services/search/test/node/search.test.ts
@@ -10,7 +10,7 @@ import { join, normalize } from 'vs/base/common/paths';
 import * as platform from 'vs/base/common/platform';
 import { joinPath } from 'vs/base/common/resources';
 import { URI } from 'vs/base/common/uri';
-import { IFolderQuery, QueryType } from 'vs/platform/search/common/search';
+import { IFolderQuery, QueryType } from 'vs/workbench/services/search/common/search';
 import { Engine as FileSearchEngine, FileWalker } from 'vs/workbench/services/search/node/fileSearch';
 import { IRawFileMatch } from 'vs/workbench/services/search/node/search';
